refactors junos libs to add support for new connection plugins
* adds junos_local lib for working with local connections
* updates junos lib for working with network_cli and netconf plugins
